Fix IndexOutOfRangeException when all children are invisible
Currently there is a crash when all children are insivible at the time of the first paint. 
But i consider it a valid usecase to have all possible elements in the UniformItemsLayout, show only the ones you enabled in your ViewModel and let the Layout figure out the presentation. 

This only adds a small check to skip the measuring when no visible children are present.